Extracr/requirements.txt
pillow
pytesseract
opencv-python
psycopg2-binary
python-dotenv
